+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. The format 
 ### Added
 - Add `project(mean=True)` to project cell mean-values to mesh-points. Now `project()` supports Triangles and Tetrahedrons.
 - Add `RegionBoundary.mesh_faces()` for a mesh with face-cells on the selected boundary of a region.
+- Add pseudo-elastic material `OgdenRoxburgh()` which may be used with the new `SolidBodyTensorNearlyIncompressible()`.
 
 ## [5.3.1] - 2022-11-03